Very rough initial -ftime-trace support
Adds an -ftime-trace flag to `zig` and `zig build`, this will now result
in clang's -ftime-trace functionality being invoked for when it is
compiling code - but a single file is generated at the output (instead
of the file being put relative to the artifacts in the desired output
folder) - so this needs fixing.

Additionally, I took a "throw things to the wall and see what sticks"
approach to adding support for this argument - next step will be
properly identifying what is needed to get this to work for C/C++ only
in the desired way - and then to ensure we can do the same for zig
generated artifacts (including proper comptime evaluation support!).